tidy up ui tests
