+++ app/pages/sections/description/ui-tests/general.ui.test.js
@@ -1,0 +1,109 @@
+import nock from 'nock';
+import { ClientFunction, Selector } from 'testcafe';
+import { extractInnerText } from 'buying-catalogue-library';
+import content from '../manifest.json';
+import { orderApiUrl } from '../../../../config';
+
+const pageUrl = 'http://localhost:1234/order/organisation/order-id/description';
+
+const setCookies = ClientFunction(() => {
+  const cookieValue = JSON.stringify({
+    id: '88421113', name: 'Cool Dude', ordering: 'manage', primaryOrganisationId: 'org-id',
+  });
+
+  document.cookie = `fakeToken=${cookieValue}`;
+});
+
+const mocks = () => {
+  nock(orderApiUrl)
+    .get('/api/v1/orders/order-id/sections/description')
+    .reply(200, { description: 'a lovely description' });
+};
+
+const pageSetup = async (withAuth = true) => {
+  if (withAuth) {
+    mocks();
+    await setCookies();
+  }
+};
+
+const getLocation = ClientFunction(() => document.location.href);
+
+fixture('Description page (existing order)')
+  .page('http://localhost:1234/order/some-fake-page')
+  .afterEach(async (t) => {
+    const isDone = nock.isDone();
+    if (!isDone) {
+      nock.cleanAll();
+    }
+
+    await t.expect(isDone).ok('Not all nock interceptors were used!');
+  });
+
+test('when user is not authenticated - should navigate to the identity server login page', async (t) => {
+  nock('http://identity-server')
+    .get('/login')
+    .reply(200);
+
+  await pageSetup(false);
+  await t.navigateTo(pageUrl);
+
+  await t
+    .expect(getLocation()).eql('http://identity-server/login');
+});
+
+test('should render description page', async (t) => {
+  await pageSetup();
+  await t.navigateTo(pageUrl);
+  const page = Selector('[data-test-id="description-page"]');
+
+  await t
+    .expect(page.exists).ok();
+});
+
+test('should render the title', async (t) => {
+  await pageSetup();
+  await t.navigateTo(pageUrl);
+
+  const title = Selector('h1[data-test-id="description-page-title"]');
+
+  await t
+    .expect(title.exists).ok()
+    .expect(await extractInnerText(title)).eql(content.title);
+});
+
+test('should render the description', async (t) => {
+  await pageSetup();
+  await t.navigateTo(pageUrl);
+
+  const description = Selector('h2[data-test-id="description-page-description"]');
+
+  await t
+    .expect(description.exists).ok()
+    .expect(await extractInnerText(description)).eql(content.description);
+});
+
+test('should render a textarea for description', async (t) => {
+  await pageSetup();
+  await t.navigateTo(pageUrl);
+
+  const description = Selector('[data-test-id="question-description"]');
+  const footerAdvice = Selector('[data-test-id="textarea-field-footer"] span');
+
+  await t
+    .expect(description.exists).ok()
+    .expect(description.find('textarea').count).eql(1)
+    .expect(footerAdvice.exists).ok()
+    .expect(await extractInnerText(footerAdvice)).eql(content.questions[0].footerAdvice);
+});
+
+test('should render save button', async (t) => {
+  await pageSetup();
+  await t.navigateTo(pageUrl);
+
+  const button = Selector('[data-test-id="save-button"] button');
+
+  await t
+    .expect(button.exists).ok()
+    .expect(await extractInnerText(button)).eql(content.saveButtonText);
+});
