+++ app/pages/sections/description/ui-tests/neworder.ui.test.js
@@ -14,10 +14,24 @@ const setCookies = ClientFunction(() => {
   document.cookie = `fakeToken=${cookieValue}`;
 });
 
-const pageSetup = async (t, withAuth = false) => {
-  if (withAuth) {
-    await setCookies();
-  }
+const postDescriptionErrorResponse = {
+  errors: [
+    {
+      field: 'Description',
+      id: 'OrderDescriptionTooLong',
+    },
+  ],
+};
+
+const mocks = () => {
+  nock(orderApiUrl)
+    .post('/api/v1/orders')
+    .reply(400, postDescriptionErrorResponse);
+};
+
+const pageSetup = async (withAuth = true, postErrorNock = false) => {
+  if (withAuth) await setCookies();
+  if (postErrorNock) mocks();
 };
 
 const getLocation = ClientFunction(() => document.location.href);
@@ -33,38 +47,8 @@ fixture('Description page (new order)')
     await t.expect(isDone).ok('Not all nock interceptors were used!');
   });
 
-const postDescriptionErrorResponse = {
-  errors: [
-    {
-      field: 'Description',
-      id: 'OrderDescriptionTooLong',
-    },
-  ],
-};
-
-test('when user is not authenticated - should navigate to the identity server login page', async (t) => {
-  await pageSetup(t);
-  nock('http://identity-server')
-    .get('/login')
-    .reply(200);
-
-  await t.navigateTo(pageUrl);
-
-  await t
-    .expect(getLocation()).eql('http://identity-server/login');
-});
-
-test('should render description page', async (t) => {
-  await pageSetup(t, true);
-  await t.navigateTo(pageUrl);
-  const page = Selector('[data-test-id="description-page"]');
-
-  await t
-    .expect(page.exists).ok();
-});
-
 test('should navigate to /organisation/neworder when click on backLink', async (t) => {
-  await pageSetup(t, true);
+  await pageSetup();
   await t.navigateTo(pageUrl);
 
   const goBackLink = Selector('[data-test-id="go-back-link"] a');
@@ -75,44 +59,8 @@ test('should navigate to /organisation/neworder when click on backLink', async (
     .expect(getLocation()).eql('http://localhost:1234/order/organisation/neworder');
 });
 
-test('should render the title', async (t) => {
-  await pageSetup(t, true);
-  await t.navigateTo(pageUrl);
-
-  const title = Selector('h1[data-test-id="description-page-title"]');
-
-  await t
-    .expect(title.exists).ok()
-    .expect(await extractInnerText(title)).eql(content.title);
-});
-
-test('should render the description', async (t) => {
-  await pageSetup(t, true);
-  await t.navigateTo(pageUrl);
-
-  const description = Selector('h2[data-test-id="description-page-description"]');
-
-  await t
-    .expect(description.exists).ok()
-    .expect(await extractInnerText(description)).eql(content.description);
-});
-
-test('should render a textarea for description', async (t) => {
-  await pageSetup(t, true);
-  await t.navigateTo(pageUrl);
-
-  const description = Selector('[data-test-id="question-description"]');
-  const footerAdvice = Selector('[data-test-id="textarea-field-footer"] span');
-
-  await t
-    .expect(description.exists).ok()
-    .expect(description.find('textarea').count).eql(1)
-    .expect(footerAdvice.exists).ok()
-    .expect(await extractInnerText(footerAdvice)).eql(content.questions[0].footerAdvice);
-});
-
 test('should not populate the text area with existing decription data', async (t) => {
-  await pageSetup(t, true);
+  await pageSetup();
   await t.navigateTo(pageUrl);
 
   const description = Selector('[data-test-id="question-description"] textarea');
@@ -122,23 +70,12 @@ test('should not populate the text area with existing decription data', async (t
     .expect(description.value).eql('');
 });
 
-test('should render save button', async (t) => {
-  await pageSetup(t, true);
-  await t.navigateTo(pageUrl);
-
-  const button = Selector('[data-test-id="save-button"] button');
-
-  await t
-    .expect(button.exists).ok()
-    .expect(await extractInnerText(button)).eql(content.saveButtonText);
-});
-
 test('should navigate to task list page when valid description is added and save is clicked', async (t) => {
   nock(orderApiUrl)
     .post('/api/v1/orders')
     .reply(200, { orderId: 'order1' });
 
-  await pageSetup(t, true);
+  await pageSetup();
   await t.navigateTo(pageUrl);
 
   const saveButton = Selector('[data-test-id="save-button"] button');
@@ -150,11 +87,7 @@ test('should navigate to task list page when valid description is added and save
 });
 
 test('should show the error summary when there are validation errors', async (t) => {
-  nock(orderApiUrl)
-    .post('/api/v1/orders')
-    .reply(400, postDescriptionErrorResponse);
-
-  await pageSetup(t, true);
+  await pageSetup(true, true);
   await t.navigateTo(pageUrl);
 
   const saveButton = Selector('[data-test-id="save-button"] button');
@@ -171,11 +104,7 @@ test('should show the error summary when there are validation errors', async (t)
 });
 
 test('should show text fields as errors with error message when there are validation errors', async (t) => {
-  nock(orderApiUrl)
-    .post('/api/v1/orders')
-    .reply(400, postDescriptionErrorResponse);
-
-  await pageSetup(t, true);
+  await pageSetup(true, true);
   await t.navigateTo(pageUrl);
 
   const descriptionPage = Selector('[data-test-id="description-page"]');
@@ -192,11 +121,7 @@ test('should show text fields as errors with error message when there are valida
 });
 
 test('should anchor to the field when clicking on the error link in errorSummary ', async (t) => {
-  nock(orderApiUrl)
-    .post('/api/v1/orders')
-    .reply(400, postDescriptionErrorResponse);
-
-  await pageSetup(t, true);
+  await pageSetup(true, true);
   await t.navigateTo(pageUrl);
 
   const saveButton = Selector('[data-test-id="save-button"] button');
