+++ app/pages/sections/service-recipients/ui-tests/general.ui.test.js
@@ -14,7 +14,6 @@ const setCookies = ClientFunction(() => {
   document.cookie = `fakeToken=${cookieValue}`;
 });
 
-
 const mocks = () => {
   nock(organisationApiUrl)
     .get('/api/v1/Organisations/org-id/service-recipients')
@@ -24,7 +23,7 @@ const mocks = () => {
     .reply(200, { serviceRecipients: [] });
 };
 
-const pageSetup = async (t, withAuth = false) => {
+const pageSetup = async (withAuth = true) => {
   if (withAuth) {
     mocks();
     await setCookies();
@@ -47,11 +46,11 @@ fixture('service-recipients page - general')
   });
 
 test('when user is not authenticated - should navigate to the identity server login page', async (t) => {
-  await pageSetup(t);
   nock('http://identity-server')
     .get('/login')
     .reply(200);
 
+  await pageSetup(false);
   await t.navigateTo(pageUrl);
 
   await t
@@ -59,7 +58,7 @@ test('when user is not authenticated - should navigate to the identity server lo
 });
 
 test('should render service-recipients page', async (t) => {
-  await pageSetup(t, true);
+  await pageSetup();
   await t.navigateTo(pageUrl);
   const page = Selector('[data-test-id="service-recipients-page"]');
 
@@ -68,7 +67,7 @@ test('should render service-recipients page', async (t) => {
 });
 
 test('should navigate to /organisation/order-id when click on backLink', async (t) => {
-  await pageSetup(t, true);
+  await pageSetup();
   await t.navigateTo(pageUrl);
 
   const goBackLink = Selector('[data-test-id="go-back-link"] a');
@@ -80,7 +79,7 @@ test('should navigate to /organisation/order-id when click on backLink', async (
 });
 
 test('should render the title', async (t) => {
-  await pageSetup(t, true);
+  await pageSetup();
   await t.navigateTo(pageUrl);
 
   const title = Selector('h1[data-test-id="service-recipients-page-title"]');
@@ -91,7 +90,7 @@ test('should render the title', async (t) => {
 });
 
 test('should render the description', async (t) => {
-  await pageSetup(t, true);
+  await pageSetup();
   await t.navigateTo(pageUrl);
 
   const description = Selector('h2[data-test-id="service-recipients-page-description"]');
@@ -102,7 +101,7 @@ test('should render the description', async (t) => {
 });
 
 test('should render the inset advice', async (t) => {
-  await pageSetup(t, true);
+  await pageSetup();
   await t.navigateTo(pageUrl);
 
   const insetAdvice = Selector('[data-test-id="service-recipients-page-insetAdvice"]');
@@ -113,7 +112,7 @@ test('should render the inset advice', async (t) => {
 });
 
 test('should render the "Select/Deselect" button', async (t) => {
-  await pageSetup(t, true);
+  await pageSetup();
   await t.navigateTo(pageUrl);
 
   const button = Selector('[data-test-id="select-deselect-button"] button');
@@ -124,7 +123,7 @@ test('should render the "Select/Deselect" button', async (t) => {
 });
 
 test('should render the organisation heading', async (t) => {
-  await pageSetup(t, true);
+  await pageSetup();
   await t.navigateTo(pageUrl);
 
   const heading = Selector('div[data-test-id="organisation-heading"]');
@@ -135,7 +134,7 @@ test('should render the organisation heading', async (t) => {
 });
 
 test('should render the ods code heading', async (t) => {
-  await pageSetup(t, true);
+  await pageSetup();
   await t.navigateTo(pageUrl);
 
   const heading = Selector('div[data-test-id="ods-code-heading"]');
@@ -146,7 +145,7 @@ test('should render the ods code heading', async (t) => {
 });
 
 test('should render the "Continue" button', async (t) => {
-  await pageSetup(t, true);
+  await pageSetup();
   await t.navigateTo(pageUrl);
 
   const button = Selector('[data-test-id="continue-button"] button');
