+++ testcafeRunner.js
@@ -11,15 +11,14 @@ const app = new App(authProvider).createApp();
 app.use(baseUrl, routes(authProvider, fakeSessionManager()));
 const server = app.listen('1234');
 
-const browserFromArgs = process.argv.slice(2, 3);
-const browserToRun = browserFromArgs.length > 0 ? browserFromArgs : 'chrome:headless';
+const [,, browserFromArgs, folderFromArgs, fileFromArgs] = process.argv;
 
-const testFromArgs = process.argv.slice(3, 4);
-const testsToRun = testFromArgs ? `**/*${testFromArgs}*/ui-tests/*ui.test.js` : '**/*ui.test.js';
+const browserToRun = browserFromArgs || 'chrome:headless';
+const testsToRun = `**/${folderFromArgs || '**'}/ui-tests/${fileFromArgs ? `${fileFromArgs}.` : '*'}ui.test.js`;
 
 let stopOnFirstFail = true;
 let quarantineMode = true;
-if (env === 'pipeline' || browserFromArgs.length > 0) {
+if (env === 'pipeline' || browserFromArgs) {
   stopOnFirstFail = false;
   quarantineMode = false;
 }
